requirements.txt
PyQt6>=6.5
argon2-cffi>=21.3.0
cryptography>=41.0
openpyxl>=3.1
python-docx>=0.8.11
reportlab>=3.6.13
validators>=0.20
